% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get Path to Package Cache Directory}
\usage{
get_cache_dir()
}
\value{
A character string path
}
\description{
Optionally overridden with HOWTHEYVOTER_CACHE_DIR env variable.
}
