% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{htv_delete_oldest_cached}
\alias{htv_delete_oldest_cached}
\title{Delete N Oldest Cached Tags (by Sorted Name)}
\usage{
htv_delete_oldest_cached(n = 2, path = get_cache_dir())
}
\arguments{
\item{n}{Number of oldest tags to delete}

\item{path}{Cache base directory}
}
\value{
Invisibly returns deleted paths
}
\description{
Delete N Oldest Cached Tags (by Sorted Name)
}
