% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htv_get_data.R
\name{htv_get_data}
\alias{htv_get_data}
\title{Download and Unzip GitHub Release Asset (if needed)}
\usage{
htv_get_data(
  repo = "HowTheyVote/data",
  tag = "latest",
  dest = get_cache_dir(),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{repo}{GitHub repository in the format "owner/repo"}

\item{tag}{GitHub release tag. Defaults to "latest"}

\item{dest}{Directory to store extracted contents. Defaults to user cache dir.}

\item{overwrite}{Logical. Should existing files be overwritten?}

\item{...}{Additional arguments passed to \code{pb_download()}}
}
\value{
A character vector of extracted file paths
}
\description{
Download and Unzip GitHub Release Asset (if needed)
}
