% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_list_metadata.R
\name{htv_list_metadata}
\alias{htv_list_metadata}
\title{Get Metadata Summary for Latest Tag}
\usage{
htv_list_metadata(
  repo = "HowTheyVote/data",
  dest = get_cache_dir(),
  tag = "latest"
)
}
\arguments{
\item{repo}{GitHub repository in the format "owner/repo"}

\item{dest}{Base destination directory}
}
\value{
A tibble with one row per CSV file and nested metadata for each column
}
\description{
Get Metadata Summary for Latest Tag
}
