% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htv_read_data.R
\name{htv_read_data}
\alias{htv_read_data}
\title{Read a CSV from cached release}
\usage{
htv_read_data(
  file,
  repo = "HowTheyVote/data",
  tag = "latest",
  dest = get_cache_dir(),
  ...
)
}
\arguments{
\item{file}{CSV filename (e.g. "data.csv")}

\item{repo}{GitHub repo "owner/repo"}

\item{tag}{Release tag ("latest" by default)}

\item{dest}{Cache dir (defaults via get_cache_dir())}

\item{...}{Passed to readr::read_csv}
}
\value{
A tibble or data.frame
}
\description{
Read a CSV from cached release
}
