% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_release_tag}
\alias{get_release_tag}
\title{Get the Latest Release Tag from GitHub}
\usage{
get_release_tag(repo = "HowTheyVote/data")
}
\arguments{
\item{repo}{GitHub repository in the format "owner/repo"}
}
\value{
A character string with the latest release tag
}
\description{
Get the Latest Release Tag from GitHub
}
