% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htv_read_data.R
\name{htv_read_data}
\alias{htv_read_data}
\title{Read a CSV from Cached GitHub Release}
\usage{
htv_read_data(
  file,
  repo = "HowTheyVote/data",
  tag = "latest",
  dest = get_cache_dir(),
  verbose = F,
  ...
)
}
\arguments{
\item{file}{CSV filename (e.g. \code{"data.csv"}).}

\item{repo}{GitHub repository in \code{"owner/repo"} format. Default is \code{"HowTheyVote/data"}.}

\item{tag}{Release tag to read from. Defaults to \code{"latest"}.}

\item{dest}{Path to the cache directory. Defaults to \code{get_cache_dir()}.}

\item{verbose}{Logical; if \code{TRUE}, shows column type messages from \code{readr::read_csv()}.}

\item{...}{Additional arguments passed to \code{readr::read_csv()}.}
}
\value{
A tibble or data.frame read from the specified CSV.
}
\description{
This function ensures that the specified release tag is downloaded and unzipped
into the local cache directory. It then reads a CSV file from that cached data.
}
